Listen Think Speak
A deconstruction of the 3 acts that put together the concept of interactivity
Using two additive synthesisers and the jitter matrix, I am making a musical conversation with the computer that is being brought alive in the jit window.
Listening act is made through the note input and the yellow gradient
Thinking act through the random scale note generator and the switching of the colors
Speaking act through the note feedback and blue colour.